+++ src/common_clj/component/kafka/adapters.clj
@@ -3,7 +3,7 @@
             [common-clj.component.kafka.models :as component.kafka.models]
             [schema.core :as s]))
 
-(s/defn kafka-record->clj-message :- component.kafka.models/KafkaMessage
+(s/defn ^:deprecated kafka-record->clj-message :- component.kafka.models/KafkaMessage
   [record]
   (let [message (json/decode (.value record) true)]
     {:topic (keyword (.topic record))